backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
python-multipart==0.0.20
pydantic-settings==2.10.1
python-dotenv==1.1.1
openai==2.15.0
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0
